@@ -1,25 +1,28 @@
-import 'package:calculadora/controllers/calculator_controller.dart';
-import 'package:calculadora/widgets/line_separator.dart';
-import 'package:calculadora/widgets/main_result.dart';
-import 'package:calculadora/widgets/sub_result.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
+import '../controllers/calculator_controller.dart';
+import 'line_separator.dart';
+import 'main_result.dart';
+import 'sub_result.dart';
+
 class MathResult extends StatelessWidget {
-  MathResult({Key? key}) : super(key: key);
+  MathResult({super.key});
 
   final calculatorCtrl = Get.find<CalculatorController>();
 
   @override
   Widget build(BuildContext context) {
-    return Obx(() => Column(
-          children: [
-            SubResult(text: '${calculatorCtrl.firstNumber}'),
-            SubResult(text: '${calculatorCtrl.operation}'),
-            SubResult(text: '${calculatorCtrl.secondNumber}'),
-            LineSeparator(),
-            MainResultText(text: '${calculatorCtrl.mathResult}'),
-          ],
-        ));
+    return Obx(
+      () => Column(
+        children: [
+          SubResult(text: '${calculatorCtrl.firstNumber}'),
+          SubResult(text: '${calculatorCtrl.operation}'),
+          SubResult(text: '${calculatorCtrl.secondNumber}'),
+          LineSeparator(),
+          MainResultText(text: '${calculatorCtrl.mathResult}'),
+        ],
+      ),
+    );
   }
 }
