@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
 
 class SubResult extends StatelessWidget {
-
   final String text;
 
-  const SubResult({Key? key, required this.text}) : super(key: key);
+  const SubResult({super.key, required this.text});
 
   @override
   Widget build(BuildContext context) {
     return Container(
       width: double.infinity,
       alignment: Alignment.centerRight,
-      child: Text( this.text , style: TextStyle(fontSize: 30 ) ),
+      child: Text(text, style: TextStyle(fontSize: 30)),
     );
   }
 }
