@@ -1,9 +1,12 @@
-import 'package:calculadora/screens/calculator_screen.dart';
 import 'package:flutter/material.dart';
+
+import 'screens/calculator_screen.dart';
 
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
